requirements.txt
gunicorn~=20.1.0
Flask~=2.0.2
Flask-RESTful~=0.3.9
Flask-SQLAlchemy~=2.5.1
Werkzeug~=2.0.2
SQLAlchemy~=1.4.27
psycopg2~=2.9.2
flask_bcrypt~=0.7.1
PyJWT~=1.7.1
Flask-Cors~=3.0.10
